Fix memory safety issues in interpolation, Kernel, and BRIEF
- Invert bounds checks in interpolate_bilinear and interpolate_bicubic to
  use positive logic, correctly rejecting NaN coordinates that would bypass
  negative bounds checks and cause out-of-bounds reads.
- Widen Kernel::new dimension check to u64 to prevent u32 overflow from
  bypassing the assertion, allowing mismatched kernel dimensions.
- Validate user-supplied BRIEF test pair coordinates against patch bounds
  to prevent out-of-bounds reads, and harden local_pixel_average against
  u32 overflow using saturating arithmetic.

Bump version to 0.26.2.

+++ src/kernel.rs
@@ -23,7 +23,8 @@ where
     /// 2. If `width * height != data.len() as u32`.
     pub const fn new(data: &'a [K], width: u32, height: u32) -> Kernel<'a, K> {
         assert!(width > 0 && height > 0, "width and height must be non-zero");
-        assert!(width * height == data.len() as u32);
+        // Widen to u64 to prevent u32 overflow from bypassing the assertion.
+        assert!(width as u64 * height as u64 == data.len() as u64);
         Kernel {
             data,
             width,
